fix(phpstan): pin remaining type mismatches in MetricsWiring, SafePath, StickinessContext
- MetricsWiring: pass \ of the Unauthorized enum case to
  Response::text (was passing the enum case itself; Response::text
  expects an int status).
- StickinessContext::markWrite: cast hrtime(true) to int before
  arithmetic so the WeakMap slot pinExpiresAtNs stays int|null and
  not int|float|null.
- SafePath: relax the constructor's $absolute / $boundary PHPDoc
  types from non-empty-string to string. The verifyUnderBoundary()
  factory is the only caller and produces non-empty paths in
  practice, but PHPStan cannot prove the non-empty property from
  the upstream realpath return shape, and the value object already
  validates traversal/symlink escapes.

+++ src/Database/Routing/StickinessContext.php
@@ -70,7 +70,7 @@ final class StickinessContext
 
         $this->slots[$key] = [
             'writeOccurred' => true,
-            'pinExpiresAtNs' => hrtime(true) + ($durationMs * 1_000_000),
+            'pinExpiresAtNs' => (int) hrtime(true) + ($durationMs * 1_000_000),
             'requestScoped' => false,
         ];
     }
